+++ lib/fog/azurerm/requests/network/list_network_security_rules.rb
@@ -17,6 +17,35 @@ module Fog
           nsr_list_result.value
         end
       end
+
+      # Mock class for Network Security Rule Request
+      class Mock
+        def list_network_security_rules(*)
+          nsr_list = '{
+            "value":[
+              {
+                "name":"myNsRule",
+                "id":"/subscriptions/{guid}/resourceGroups/{resourceGroupName}/providers/Microsoft.Network/networkSecurityGroups/myNsg/securityRules/myNsRule",
+                "etag":"W/\"00000000-0000-0000-0000-000000000000\"",
+                "properties":{
+                  "provisioningState":"Succeeded",
+                  "description":"description-of-this-rule",
+                  "protocol": "*",
+                  "sourcePortRange":"source-port-range",
+                  "destinationPortRange":"destination-port-range",
+                  "sourceAddressPrefix":"*",
+                  "destinationAddressPrefix":"*",
+                  "access":"Allow",
+                  "priority":100,
+                  "direction":"Inbound"
+                }
+              }
+          ]
+          }'
+          nsr_mapper = Azure::ARM::Network::Models::SecurityRuleListResult.mapper
+          @network_client.deserialize(nsr_mapper, JSON.load(nsr_list), 'result.body')
+        end
+      end
     end
   end
 end